feat: implement muscle/fat segmentation feature with mask generation
- Added a new endpoint to segment a volume into a binary muscle/fat mask.
- Introduced a MaskDialog component to preview the segmentation results.
- Updated the API client to handle mask segmentation requests and responses.
- Enhanced volume loading logic to prevent duplicate uploads of the same file.
- Refactored volume index handling to use Uint32Array for better performance with large datasets.
- Added tests for the new mask endpoint and various processing functions.
- Improved error handling and state management in stitching and file loading processes.

+++ frontend/src/shared/h5/volumeCache.ts
@@ -22,7 +22,7 @@ interface CachedVolume {
     nSlices: number
     height: number
     width: number
-    vIndices: Float32Array
+    vIndices: Uint32Array
     vIntensities: Float32Array
     normalizedVolume: Uint8Array | null
 }
@@ -77,7 +77,7 @@ function isCachedVolume(v: unknown): v is CachedVolume {
         v !== null &&
         typeof v === 'object' &&
         typeof (v as CachedVolume).nSlices === 'number' &&
-        (v as CachedVolume).vIndices instanceof Float32Array &&
+        (v as CachedVolume).vIndices instanceof Uint32Array &&
         (v as CachedVolume).vIntensities instanceof Float32Array
     )
 }
